added NIOLockedBox over histogram and requestResult count
