+++ Sources/performance-worker/BenchmarkClient.swift
@@ -15,63 +15,61 @@
  */
 
 import GRPCCore
+import NIOConcurrencyHelpers
 
 @available(macOS 13.0, iOS 16.0, watchOS 9.0, tvOS 16.0, *)
 struct BenchmarkClient {
   var client: GRPCClient
   var rpcNumber: Int32
   var rpcType: Grpc_Testing_RpcType
-  var histogramParams: Grpc_Testing_HistogramParams
+  private let rpcStats: NIOLockedValueBox<RPCStats>
 
   init(
     client: GRPCClient,
     rpcNumber: Int32,
     rpcType: Grpc_Testing_RpcType,
-    histogramParams: Grpc_Testing_HistogramParams
+    histogramParams: Grpc_Testing_HistogramParams?
   ) {
     self.client = client
     self.rpcNumber = rpcNumber
     self.rpcType = rpcType
-    self.histogramParams = histogramParams
+
+    var histogram: RPCStats.LatencyHistogram
+    if let histogramParams = histogramParams {
+      histogram = .init(
+        resolution: histogramParams.resolution,
+        maxBucketStart: histogramParams.maxPossible
+      )
+    } else {
+      histogram = .init()
+    }
+
+    self.rpcStats = NIOLockedValueBox(RPCStats(latencyHistogram: histogram))
   }
 
-  func run() async throws -> LatencyHistogram {
+  func run() async throws {
     let benchmarkClient = Grpc_Testing_BenchmarkServiceClient(client: client)
-    return try await withThrowingTaskGroup(of: Void.self, returning: LatencyHistogram.self) {
-      clientGroup in
+    return try await withThrowingTaskGroup(of: Void.self) { clientGroup in
       // Start the client.
       clientGroup.addTask { try await client.run() }
 
       // Make the requests to the server and register the latency for each one.
-      let latencies = try await withThrowingTaskGroup(of: Double.self, returning: [Double].self) {
-        rpcsGroup in
+      try await withThrowingTaskGroup(of: Void.self) { rpcsGroup in
         for _ in 0 ..< self.rpcNumber {
           rpcsGroup.addTask {
-            return self.makeRPC(client: benchmarkClient, rpcType: self.rpcType)
+            let (latency, errorCode) = self.makeRPC(client: benchmarkClient, rpcType: self.rpcType)
+            self.rpcStats.withLockedValue {
+              $0.latencyHistogram.add(value: latency)
+              if errorCode != nil {
+                $0.requestResultCount[Int32(errorCode!.rawValue), default: 1] += 1
+              }
+            }
           }
         }
-
-        var latencies = [Double]()
-
-        while let latency = try await rpcsGroup.next() {
-          latencies.append(latency)
-        }
-
-        return latencies
+        try await rpcsGroup.waitForAll()
       }
 
       try await clientGroup.next()
-
-      // Creating the LatencyHistogram for the current client.
-      var latencyLatencyHistogram = LatencyHistogram(
-        resolution: self.histogramParams.resolution,
-        maxBucketStart: self.histogramParams.maxPossible
-      )
-      for latency in latencies {
-        latencyLatencyHistogram.add(value: latency)
-      }
-
-      return latencyLatencyHistogram
     }
   }
 
@@ -79,13 +77,15 @@ struct BenchmarkClient {
   private func makeRPC(
     client: Grpc_Testing_BenchmarkServiceClient,
     rpcType: Grpc_Testing_RpcType
-  ) -> Double {
+  ) -> (latency: Double, errorCode: RPCError.Code?) {
     switch rpcType {
     case .unary, .streaming, .streamingFromClient, .streamingFromServer, .streamingBothWays,
       .UNRECOGNIZED:
-      let startTime = LatencyHistogram.grpcTimeNow()
-      let endTime = LatencyHistogram.grpcTimeNow()
-      return Double((endTime - startTime).value)
+      let startTime = RPCStats.LatencyHistogram.grpcTimeNow()
+      let endTime = RPCStats.LatencyHistogram.grpcTimeNow()
+      return (
+        latency: Double((endTime - startTime).value), errorCode: RPCError.Code(.unimplemented)
+      )
     }
   }
 }
